@@ -16,9 +16,8 @@
     "@next/bundle-analyzer": "^13.2.4",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/forms": "^0.5.3",
-    "@tailwindcss/line-clamp": "^0.4.2",
     "@tailwindcss/typography": "^0.5.9",
-    "@typescript-eslint/eslint-plugin": "^5.56.0",
+    "@typescript-eslint/eslint-plugin": "^5.57.1",
     "autoprefixer": "10.4.14",
     "bson-objectid": "^2.0.4",
     "date-fns": "^2.29.3",
@@ -26,33 +25,32 @@
     "fast-deep-equal": "^3.1.3",
     "idb": "^7.1.1",
     "jest": "^29.5.0",
-    "jotai": "^2.0.3",
-    "mongoose": "^7.0.2",
+    "mongoose": "^7.0.3",
     "next": "^13.2.4",
-    "next-auth": "^4.20.1",
+    "next-auth": "^4.21.1",
     "next-router-mock": "^0.9.3",
     "nodemailer": "^6.9.1",
     "react": "^18.2.0",
     "react-device-detect": "^2.2.3",
     "react-dom": "^18.2.0",
-    "react-error-boundary": "^4.0.2",
+    "react-error-boundary": "^4.0.3",
     "react-popper-tooltip": "^4.4.2",
     "recoil": "^0.7.7",
-    "sharp": "^0.31.3",
-    "slate": "^0.91.4",
-    "slate-history": "^0.86.0",
+    "sharp": "^0.32.0",
+    "slate": "^0.93.0",
+    "slate-history": "^0.93.0",
     "slate-hyperscript": "^0.77.0",
-    "slate-react": "^0.92.0",
+    "slate-react": "^0.93.0",
     "validator": "^13.9.0"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
     "@types/jest": "^29.5.0",
-    "@types/node": "18.15.5",
-    "@types/react": "^18.0.28",
+    "@types/node": "18.15.11",
+    "@types/react": "^18.0.33",
     "@types/validator": "^13.7.14",
-    "eslint": "8.36.0",
+    "eslint": "8.37.0",
     "eslint-config-next": "13.2.4",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-testing-library": "^5.10.2",
@@ -60,9 +58,9 @@
     "jest-environment-jsdom": "^29.5.0",
     "lint-staged": "^13.2.0",
     "postcss": "^8.4.21",
-    "prettier": "^2.8.6",
-    "prettier-plugin-tailwindcss": "^0.2.5",
-    "tailwindcss": "^3.2.7",
-    "typescript": "5.0.2"
+    "prettier": "^2.8.7",
+    "prettier-plugin-tailwindcss": "^0.2.7",
+    "tailwindcss": "^3.3.1",
+    "typescript": "5.0.3"
   }
 }
