@@ -1,33 +1,9 @@
-import { CATCH } from '@constAssertions/misc';
-import { LayoutTypeEffect } from '@effects/layoutTypeEffect';
-import { NavigationInitialEffect } from '@effects/navigationInitialEffect';
 import { Types } from '@lib/types';
-import { atomCatch, atomHtmlTitleTag } from '@states/misc';
+import { atomHtmlTitleTag } from '@states/misc';
 import dynamic from 'next/dynamic';
 import Head from 'next/head';
-import {
-  Fragment as EffectFragment,
-  Fragment as HeaderFragment,
-  Fragment as LayoutAppFragment,
-  Fragment as ModalActionsFragment,
-  Suspense,
-} from 'react';
+import { Fragment as HeaderFragment, Fragment as LayoutAppFragment, Suspense } from 'react';
 import { useRecoilValue } from 'recoil';
-const CreateTodoModal = dynamic(() =>
-  import('@modals/todoModals/todoModal').then((mod) => mod.TodoModal),
-);
-const MinimizedModal = dynamic(() =>
-  import('@modals/minimizedModal').then((mod) => mod.MinimizedModal),
-);
-const Notification = dynamic(() =>
-  import('components/notifications/notification').then((mod) => mod.Notification),
-);
-const LabelModal = dynamic(() =>
-  import('@modals/labelModals/labelModal').then((mod) => mod.LabelModal),
-);
-const WindowBeforeunloadEffect = dynamic(() =>
-  import('@effects/windowBeforeunloadEffect').then((mod) => mod.WindowBeforeunloadEffect),
-);
 const FooterBody = dynamic(() =>
   import('@layouts/layoutFooter/footerBody').then((mod) => mod.FooterBody),
 );
@@ -39,6 +15,12 @@ const LayoutFooter = dynamic(
   () => import('@layouts/layoutFooter').then((mod) => mod.LayoutFooter),
   { ssr: false },
 );
+
+const LayoutAppGroupEffects = dynamic(
+  () => import('@effects/layout').then((mod) => mod.LayoutAppGroupEffects),
+  { ssr: false },
+);
+
 const User = dynamic(() => import('@layouts/layoutHeader/user').then((mod) => mod.User), {
   ssr: false,
 });
@@ -46,13 +28,12 @@ const User = dynamic(() => import('@layouts/layoutHeader/user').then((mod) => mo
 type Props = Pick<Types, 'children'>;
 
 export const LayoutApp = ({ children }: Props) => {
-  const catchTodoModal = useRecoilValue(atomCatch(CATCH.todoModal));
   const slug = useRecoilValue(atomHtmlTitleTag);
 
   return (
     <LayoutAppFragment>
       <Head>
-        <title>{'My Todo App - ' + slug}</title>
+        <title>{slug ? 'Todos - ' + slug : ''}</title>
       </Head>
       <HeaderFragment>
         <div className='flex h-screen flex-col'>
@@ -67,17 +48,7 @@ export const LayoutApp = ({ children }: Props) => {
           </LayoutFooter>
         </div>
       </HeaderFragment>
-      <EffectFragment>
-        <LayoutTypeEffect layoutType='app' />
-        <NavigationInitialEffect layoutType='app' />
-        <Notification />
-        <WindowBeforeunloadEffect />
-        <ModalActionsFragment>
-          <CreateTodoModal />
-          <MinimizedModal />
-          {!catchTodoModal && <LabelModal />}
-        </ModalActionsFragment>
-      </EffectFragment>
+      <LayoutAppGroupEffects />
     </LayoutAppFragment>
   );
 };
