@@ -11,8 +11,8 @@ import {
 import { Logo } from './logo';
 import { NavigationButton } from './navigationButton';
 
-const UserSessionResetEffect = dynamic(() =>
-  import('@effects/userSessionResetEffect').then((mod) => mod.UserSessionResetEffect),
+const UserSessionGroupEffects = dynamic(() =>
+  import('@effects/users').then((mod) => mod.UserSessionGroupEffects),
 );
 
 type Props = Partial<Pick<Types, 'children'>> & Pick<Types, 'layoutType'>;
@@ -47,17 +47,24 @@ export const LayoutHeader = ({ children, layoutType }: Props) => {
         <RightNavigationFragment>
           <div
             className={classNames(
-              layoutApp && 'mr-3 flex flex-1 flex-row items-center justify-end pl-2',
+              layoutApp && 'flex flex-1 flex-row items-center justify-end pl-2 ml:mr-3',
               layoutHome && 'hidden ml:mr-8 ml:flex',
             )}>
             {children}
           </div>
           <NavigationButtonFragment>
-            <div className='pr-6 ml:hidden ml:pr-0'>{layoutHome && <NavigationButton />}</div>
+            <div
+              className={classNames(
+                'ml:hidden ml:pr-0',
+                layoutHome && 'pr-6',
+                layoutApp && 'pr-3',
+              )}>
+              {layoutHome && <NavigationButton />}
+            </div>
           </NavigationButtonFragment>
         </RightNavigationFragment>
       </div>
-      <UserSessionResetEffect />
+      <UserSessionGroupEffects />
     </LayoutHeaderFragment>
   );
 };
