@@ -15,19 +15,28 @@ import { useConditionCheckLabelTitleEmpty } from '@hooks/misc';
 import { useLabelModalStateClose } from '@hooks/modals';
 import { ModalTransitionChild } from '@modals/modal/modalTransition/modalTransitionChild';
 import { ModalTransitionRoot } from '@modals/modal/modalTransition/modalTransitionRoot';
-import { Divider } from '@ui/dividers/divider';
 import { Button } from '@buttons/button';
 import { classNames } from '@stateLogics/utils';
 import { KeysWithLabelModalEffect } from '@effects/keysWithLabelModalEffect';
+import { DividerX } from '@ui/dividers/dividerX';
 
-type Props = Partial<Pick<Types, 'label' | 'children' | 'menuButtonContent' | 'footerButtons' | 'headerButtons'>>;
+type Props = Partial<
+  Pick<Types, 'label' | 'children' | 'menuButtonContent' | 'footerButtons' | 'headerButtons'>
+>;
 
-export const LabelModal = ({ label, footerButtons, children, menuButtonContent = 'Create new label' }: Props) => {
+export const LabelModal = ({
+  label,
+  footerButtons,
+  children,
+  menuButtonContent = 'Create new label',
+}: Props) => {
   const isLabelModalOpen = useRecoilValue(atomLabelModalOpen(label?._id));
   const closeModal = useLabelModalStateClose(label?._id);
   const initialFocusInput = useRef<HTMLInputElement>(null);
   const labelItem =
-    typeof label === 'undefined' ? useRecoilValue(atomLabelNew) : useRecoilValue(atomSelectorLabelItem(label._id));
+    typeof label === 'undefined'
+      ? useRecoilValue(atomLabelNew)
+      : useRecoilValue(atomSelectorLabelItem(label._id));
   const updateLabelItem = useLabelValueUpdate(label);
   const addLabel = useLabelAdd();
   const condition = useConditionCheckLabelTitleEmpty();
@@ -48,7 +57,7 @@ export const LabelModal = ({ label, footerButtons, children, menuButtonContent =
               />
             </div>
             <div className='hidden sm:mb-3 sm:block'>
-              <Divider />
+              <DividerX />
             </div>
           </div>
           <div className='h-full w-full overflow-scroll pl-2 pr-3'>
