@@ -3,7 +3,6 @@ import {
   IconButton as MaxIconButton,
   IconButton as OpenFullIconButton,
 } from '@buttons/iconButton';
-import { atomMediaQuery } from '@states/misc';
 import { atomTodoModalMini } from '@states/modals';
 import { atomTodoNew } from '@states/todos';
 import { TypesTodo } from 'lib/types';
@@ -18,7 +17,12 @@ import {
 } from '@options/button';
 import { optionsMinimizedModal } from '@options/misc';
 import { ModalStateOnBreakpointEffect } from '@effects/modalStateOnBreakpointEffect';
-import { useTodoModalStateClose, useTodoModalStateMaximize, useTodoModalStateExitMinimize } from '@hooks/modals';
+import {
+  useTodoModalStateClose,
+  useTodoModalStateMaximize,
+  useTodoModalStateExitMinimize,
+} from '@hooks/modals';
+import { atomEffectMediaQuery } from '@states/atomEffects/misc';
 
 type Props = Partial<Pick<TypesTodo, 'todo'>>;
 
@@ -28,7 +32,7 @@ export const MinimizedModal = ({ todo }: Props) => {
   const maximizeModal = useTodoModalStateMaximize(todo?._id);
   const exitMinimizeModal = useTodoModalStateExitMinimize(todo?._id);
   const newTodo = useRecoilValue(atomTodoNew);
-  const isMediaQuerySmall = useRecoilValue(atomMediaQuery(BREAKPOINT['sm']));
+  const isMediaQuerySmall = useRecoilValue(atomEffectMediaQuery(BREAKPOINT['sm']));
 
   return (
     <Fragment>
@@ -38,7 +42,9 @@ export const MinimizedModal = ({ todo }: Props) => {
           options={optionsMinimizedModal}>
           <div className=' flex flex-shrink-0 flex-row items-center justify-between'>
             <div className='flex flex-1 flex-col justify-center'>
-              <p className='w-44 break-words text-sm font-medium text-gray-500 line-clamp-1'>{newTodo.title}</p>
+              <p className='line-clamp-1 w-44 break-words text-sm font-medium text-gray-500'>
+                {newTodo.title}
+              </p>
             </div>
             <div className='flex h-fit flex-row'>
               <MaxIconButton
@@ -56,7 +62,7 @@ export const MinimizedModal = ({ todo }: Props) => {
             </div>
           </div>
           <div>
-            <p className='w-64 break-words text-sm text-gray-400 line-clamp-3'>{newTodo.note}</p>
+            <p className='line-clamp-3 w-64 break-words text-sm text-gray-400'>{newTodo.note}</p>
           </div>
           <ModalStateOnBreakpointEffect />
         </MinimizeModalTransition>
