@@ -1,11 +1,6 @@
 import { PATHNAME_IMAGE } from '@constAssertions/data';
 import { CATCH } from '@constAssertions/misc';
-import { BREAKPOINT, SPINNER } from '@constAssertions/ui';
-import {
-  mediaQueryEffect,
-  networkStatusEffect,
-} from '@lib/stateLogics/effects/atomEffects/atomEffects';
-import { Types } from '@lib/types';
+import { SPINNER } from '@constAssertions/ui';
 import { atom, atomFamily, selector } from 'recoil';
 
 /*
@@ -20,20 +15,6 @@ export const atomCatch = atomFamily<boolean, CATCH>({
 export const atomDisableScroll = atom<boolean>({
   key: 'atomDisableScroll',
   default: false,
-});
-
-// Media Queries
-export const atomMediaQuery = atomFamily<boolean, BREAKPOINT>({
-  key: 'atomMediaQuery',
-  default: false,
-  effects: (breakpoint) => [mediaQueryEffect({ breakpoint: breakpoint })],
-});
-
-// Network
-export const atomNetworkStatusEffect = atom({
-  key: 'atomNetworkStatusEffect',
-  default: true,
-  effects: [networkStatusEffect],
 });
 
 //  Dropdown
@@ -57,11 +38,6 @@ export const atomHtmlTitleTag = atom<string>({
   default: '',
 });
 
-export const atomPathname = atom<Types['pathname']>({
-  key: 'atomPathname',
-  default: '/',
-});
-
 export const atomPathnameImage = atom<PATHNAME_IMAGE>({
   key: 'atomPathnameImage',
   default: PATHNAME_IMAGE['focus'],
@@ -70,6 +46,11 @@ export const atomPathnameImage = atom<PATHNAME_IMAGE>({
 export const atomLoadingSpinner = atomFamily<boolean, SPINNER>({
   key: 'atomLoadingSpinner',
   default: false,
+});
+
+export const atomFilterEffect = atom({
+  key: 'atomFilterEffect',
+  default: 'focus',
 });
 
 /**
